backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Agentic backend with Crew AI integration"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "crewai>=0.18.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.11.0",
    "langchain>=0.0.331",
    "langchain-community>=0.0.10",  # For Ollama integration
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.5",
    "email-validator (>=2.2.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest]
testpaths = ["tests"]
